+++ src/main/java/com/rev/app/controller/WishlistController.java
@@ -3,16 +3,14 @@ package com.rev.app.controller;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.*;
 
 import com.rev.app.service.IWishlistService;
 
-
 import jakarta.servlet.http.HttpSession;
+import lombok.extern.slf4j.Slf4j;
 
+@Slf4j
 @Controller
 @RequestMapping("/wishlist")
 public class WishlistController {
@@ -20,21 +18,30 @@ public class WishlistController {
     @Autowired
     private IWishlistService wishlistService;
 
+    
     @PostMapping("/add/{productId}")
     public String addToWishlist(@PathVariable Long productId,
                                 HttpSession session) {
+        log.info("UI: Adding product {} to wishlist", productId);
 
         Long userId = (Long) session.getAttribute("userId");
+
+        if (userId == null)
+            return "redirect:/login";
 
         wishlistService.addToWishlist(userId, productId);
 
         return "redirect:/products";
     }
 
+   
     @GetMapping
     public String viewWishlist(HttpSession session, Model model) {
 
         Long userId = (Long) session.getAttribute("userId");
+
+        if (userId == null)
+            return "redirect:/login";
 
         model.addAttribute("wishlist",
                 wishlistService.getWishlist(userId));
@@ -42,8 +49,15 @@ public class WishlistController {
         return "wishlist";
     }
 
+    
     @GetMapping("/remove/{id}")
-    public String remove(@PathVariable Long id) {
+    public String remove(@PathVariable Long id,
+                         HttpSession session) {
+
+        Long userId = (Long) session.getAttribute("userId");
+
+        if (userId == null)
+            return "redirect:/login";
 
         wishlistService.removeFromWishlist(id);
         return "redirect:/wishlist";
